pyproject.toml
[tool.poetry]
name = "diploma-scripts"
version = "0.1.0"
description = ""
authors = ["Igor Kalinichenko <igorr.kalina@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.10"
boto3 = "^1.35.68"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
nltk = {extras = ["all"], version = "^3.9.1"}
numpy = "^2.1.3"
matplotlib = "^3.9.3"
scikit-learn = "^1.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
